Refactor: Update CLI structure by removing index.cjs, consolidating logic into run.ts, and adjusting package.json for new bin path
