+++ code/lib/cli-storybook/package.json
@@ -24,12 +24,11 @@
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
+      "default": "./dist/index.js"
     },
-    "./bin/index.cjs": "./bin/index.cjs",
     "./package.json": "./package.json"
   },
-  "bin": "./bin/index.cjs",
+  "bin": "./dist/bin/index.js",
   "files": [
     "bin/**/*",
     "dist/**/*",
@@ -37,8 +36,8 @@
     "!src/**/*"
   ],
   "scripts": {
-    "check": "jiti  ../../../scripts/prepare/check.ts",
-    "prep": "jiti  ../../../scripts/prepare/bundle.ts"
+    "check": "jiti ../../../scripts/prepare/check.ts",
+    "prep": "jiti ../../../scripts/build/index.ts"
   },
   "dependencies": {
     "@storybook/codemod": "workspace:*",
@@ -71,16 +70,6 @@
   },
   "publishConfig": {
     "access": "public"
-  },
-  "bundler": {
-    "entries": [
-      "./src/index.ts",
-      "./src/bin/index.ts"
-    ],
-    "formats": [
-      "cjs"
-    ],
-    "platform": "node"
   },
   "gitHead": "e6a7fd8a655c69780bc20b9749c2699e44beae16"
 }
